Murilo

+++ include/Entidades/Personagens/Nevoso.h
@@ -17,7 +17,7 @@ namespace Alaska
                     Nevoso(float vx, float vy, Jogador* pJ);
                     Nevoso();
                     ~Nevoso();
-                    float calcularAcumulacao();
+                    int calcularAcumulacao();
                     void danificar(Alaska::Entidades::Personagens::Jogador* pJ);
                     void executar();
                     void salvar();
